\docType{package}
\name{irisws-package}
\alias{irisws-package}
\title{Access to IRIS Web Services within R}
\description{
  Description
}
\details{
  Details
}
\references{
  \url{http://service.iris.edu/irisws}
}
\seealso{
  \code{\link{constructor}} to construct queries and
  \code{\link{query.iris}} to query IRIS WS
}

