\name{irisws-wadl}
\alias{irisws-wadl}
\alias{waddler}
\alias{wadl}
\title{Find the WADL file for an IRIS WS}
\usage{
waddler(service, ...)

wadl(service, ...)
}
\arguments{
  \item{service}{character}
}
\description{
Find the WADL file for an IRIS WS
}
\author{
A.J. Barbour
}
\references{
[1] \url{http://www.w3.org/Submission/wadl/}
}
\seealso{
\code{\link{iris.query}} to query IRIS WS

\code{\link{irisws-package}}

Other Utilities: \code{\link{ISOdatetime.j}},
\code{\link{ISOtimestring}},
\code{\link{irisws-timeutilities}},
\code{\link{timestring}}; \code{\link{check.query}},
\code{\link{iris.query}}, \code{\link{query.iris}};
\code{\link{constructor}}, \code{\link{constructor2}},
\code{\link{irisws-queryconstruction}},
\code{\link{params2queryparams}}; \code{\link{irisws}},
\code{\link{irisws-package}}
}

