\docType{package}
\name{irisws-webservices}
\alias{irisws-webservices}
\alias{irisws-webservices-package}
\alias{webservice}
\alias{webservices}
\alias{ws}
\title{Summary of webservice accessors}
\usage{
webservices()
}
\description{
Description
}
\details{
Details
}
\examples{
# list all the webservice accessors:
webservices()
}
\references{
\url{http://service.iris.edu/irisws}
}
\seealso{
\code{\link{irisws-package}}

Other WebServices: \code{\link{.llpair}},
\code{\link{PS_time.distances}},
\code{\link{PS_time.stations}}, \code{\link{traveltime}},
\code{\link{traveltimes}}, \code{\link{tt}},
\code{\link{ttDeg.ws}}, \code{\link{ttKm.ws}},
\code{\link{ttStaSrc.ws}}, \code{\link{ws.traveltime}},
\code{\link{ws.ttDeg}}, \code{\link{ws.ttDistances}},
\code{\link{ws.ttKm}}, \code{\link{ws.ttStaSrc}};
\code{\link{distaz}}, \code{\link{distaz.ws}},
\code{\link{ws.distaz}}; \code{\link{flinnengdahl}},
\code{\link{flinnengdahl.ws}},
\code{\link{ws.flinnengdahl}}; \code{\link{resp}},
\code{\link{resp.ws}}, \code{\link{ws.resp}};
\code{\link{timeseries}}, \code{\link{timeseries.ws}},
\code{\link{ws.timeseries}}
}

