\docType{package}
\name{irisws}
\alias{irisws}
\alias{irisws-package}
\title{Query IRIS Web Services through R}
\description{
  Query IRIS Web Services through R
}
\references{
  \url{http://service.iris.edu/irisws}
}
\seealso{
  \code{\link{constructor}} to construct queries and
  \code{\link{query.iris}} to query IRIS WS
}

