\name{distaz}
\alias{distaz}
\alias{distaz.ws}
\alias{ws.distaz}
\title{Access the 'distaz' Web Service for distance/azimuth computations}
\usage{
ws.distaz(station.latlon = c(0, 0), event.latlon = c(0, 0),
  filename = NULL, verbose = TRUE)

distaz.ws(station.latlon = c(0, 0), event.latlon = c(0, 0),
  filename = NULL, verbose = TRUE)
}
\arguments{
  \item{station.latlon}{numeric; the decimal latitude and
  longitude of the station}

  \item{event.latlon}{numeric; the decimal latitude and
  longitude of the event (the 'source')}

  \item{filename}{\code{NA} for an auto-generated filename
  based on the inputs; \code{NULL} for a temporary
  filename; or, a character string of the user's choosing.
  See \strong{Details} for details about the auto-generated
  name.}

  \item{verbose}{logical; should messages be given?}
}
\value{
A list (invisibly) with the query string, and data from the
result
}
\description{
From [1]: \emph{The distance-azimuth service will calculate
the great-circle angular distance, azimuth, and backazimuth
between two geographic coordinate pairs. All results are
reported in degrees, with azimuth and backazimuth measured
clockwise from North.}
}
\details{
The query is generated with \code{\link{constructor2}}, and
executed with \code{\link{query.iris}}, which throws errors
based on \code{\link{check.query}}.

\code{\link{distaz.ws}} is simply a pointer to
\code{\link{ws.distaz}}

\subsection{Filename options (\code{filename=})}{ \itemize{
\item{\code{NA}}: An auto-generated file.
\item{\code{NULL}}: A temporary file.  \item{<character
string>}: Any desired name.  } Autogenerated files are of
the form:

\code{iriswsQ.distaz}

Temporary files are generated within
\code{\link{query.iris}}. }
}
\examples{
\dontrun{

# exact same point -- should give zero
qr <- ws.distaz()
querydata(qr)

# two different locations -- note the coordinates
# must be in the order of latitude, longitude
qr <- ws.distaz(c(0.,0.),c(30.,-100.))
querydata(qr)
# should give
#   azimuth backAzimuth distance
#1 84.98686    300.2136 98.66374

}
}
\author{
AJ Barbour
}
\references{
[1] \url{http://service.iris.edu/irisws/distaz/1/}
}
\seealso{
Other WebServices: \code{\link{.llpair}},
\code{\link{PS_time.distances}},
\code{\link{PS_time.stations}}, \code{\link{traveltime}},
\code{\link{traveltimes}}, \code{\link{tt}},
\code{\link{ttDeg.ws}}, \code{\link{ttKm.ws}},
\code{\link{ttStaSrc.ws}}, \code{\link{ws.traveltime}},
\code{\link{ws.ttDeg}}, \code{\link{ws.ttDistances}},
\code{\link{ws.ttKm}}, \code{\link{ws.ttStaSrc}};
\code{\link{flinn}}, \code{\link{flinnengdahl}},
\code{\link{flinnengdahl.ws}},
\code{\link{ws.flinnengdahl}};
\code{\link{irisws-webservices}},
\code{\link{irisws-webservices-package}},
\code{\link{services}}, \code{\link{webservice}},
\code{\link{webservices}}, \code{\link{ws}};
\code{\link{resp}}, \code{\link{resp.ws}},
\code{\link{ws.resp}}; \code{\link{timeseries}},
\code{\link{timeseries.ws}}, \code{\link{ws.timeseries}}
}

