% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/webservices.R
\name{flinnengdahl}
\alias{flinn}
\alias{flinnengdahl}
\alias{flinnengdahl.ws}
\alias{ws.flinnengdahl}
\title{Access the 'flinnengdahl' Web Service for Flinn-Engdahl region information}
\usage{
ws.flinnengdahl(lat, lon, output = c("code", "region"), filename = NULL,
  verbose = FALSE)

flinnengdahl.ws(lat, lon, output = c("code", "region"), filename = NULL,
  verbose = FALSE)
}
\arguments{
\item{lat}{numeric; North latitude, in decimal degrees.
Must be within the range [-90.0, 90.0]}

\item{lon}{numeric; East longitude, in decimal degrees.
Must be within the range [-180.0, 180.0]}

\item{output}{character;
determines which Flinn-Engdahl representation is retrieved:
\code{code} retrieves the numeric code, (e.g., 32), and
\code{region} retrieves the region description, (e.g., OREGON)}

\item{filename}{\code{NA} for an auto-generated filename based on the
inputs; \code{NULL} for a temporary filename; or, a character string
of the user's choosing.  See \strong{Details} for details about
the auto-generated name.}

\item{verbose}{logical; should messages be given?}
}
\value{
A list (invisibly) with the query string, and data from the result
}
\description{
From [1]:
\emph{The Flinn-Engdahl webservice returns either the Flinn-Engdahl
region code or region name for a specified latitude and longitude.}

\emph{Returned values are text strings, e.g. region code `014'
which corresponds to region name `Kenai Peninsula, Alaska'.}
}
\details{
\subsection{Filename options (\code{filename=})}{
 \itemize{
     \item{\code{NA}}: An auto-generated file.
     \item{\code{NULL}}: A temporary file.
     \item{<character string>}: Any desired name.
 }
Autogenerated files are of the form:

\code{iriswsQ.flinnengdahl}

Temporary files are generated within \code{\link{query.iris}}.
}
}
\examples{
\dontrun{
querydata(flinnengdahl.ws(10,10)) # '755'
querydata(flinnengdahl.ws(10,10,output="region")) # 'NIGERIA'
}
}
\author{
AJ Barbour
}
\references{
[1] \url{http://service.iris.edu/irisws/flinnengdahl/2/}
}
\seealso{
Other WebServices: \code{\link{.llpair}},
  \code{\link{PS_time.distances}},
  \code{\link{PS_time.stations}}, \code{\link{traveltime}},
  \code{\link{traveltimes}}, \code{\link{tt}},
  \code{\link{ttDeg.ws}}, \code{\link{ttKm.ws}},
  \code{\link{ttStaSrc.ws}}, \code{\link{ws.traveltime}},
  \code{\link{ws.tt}}, \code{\link{ws.ttDeg}},
  \code{\link{ws.ttDistances}}, \code{\link{ws.ttKm}},
  \code{\link{ws.ttStaSrc}};
  \code{\link{Distance-Azimuth}}, \code{\link{distaz}},
  \code{\link{distaz.ws}}, \code{\link{ws.distaz}};
  \code{\link{irisws-webservices}},
  \code{\link{irisws-webservices-package}},
  \code{\link{services}}, \code{\link{webservice}},
  \code{\link{webservices}}, \code{\link{ws}};
  \code{\link{resp}}, \code{\link{resp.ws}},
  \code{\link{ws.resp}}; \code{\link{timeseries}},
  \code{\link{timeseries.ws}}, \code{\link{ws.timeseries}}
}

