% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irisws-package.R
\docType{package}
\name{irisws-package}
\alias{irisws}
\alias{irisws-package}
\title{Access to IRIS Web Services within R}
\description{
This package can be used to obtain products from the
Data Management Center  of the
Incorporated Research Institutions in Seismology [1],
through their web-services API [2].

Currenty there is support for access to
\code{\link{Distance-Azimuth}} calculations,
xxx, ...
}
\details{
The code constructs queries according to
the requirements of the API, executes them, and
optionally saves and/or loads the results into
the R environment.
A good place to start is with the \code{\link{webservices}}
function.
}
\examples{
# see the available functions
webservices()

# see which webservices are actually supported
services()
}
\author{
A.J. Barbour
}
\references{
[1] \url{http://www.iris.edu/}

[2] \url{http://service.iris.edu/irisws}
}
\seealso{
\code{\link{irisws-webservices}} to see a list
of the webservice accessors currently available
 
\code{\link{constructor}} \emph{et al.} to construct queries

\code{\link{query.iris}} to query IRIS WS

Other Utilities: \code{\link{irisws-queryconstruction}},
  \code{\link{irisws-timeutilities}},
  \code{\link{irisws-wadl}}, \code{\link{query.iris}}
}

