\docType{package}
\name{irisws-package}
\alias{irisws}
\alias{irisws-package}
\title{Access to IRIS Web Services within R}
\description{
Description
}
\details{
Details
}
\examples{
webservices()
}
\author{
A.J. Barbour
}
\references{
\url{http://service.iris.edu/irisws}
}
\seealso{
\code{\link{irisws-webservices}} to see a list of the
webservice accessors currently available

\code{\link{constructor}} \emph{et al.} to construct
queries

\code{\link{query.iris}} to query IRIS WS

Other Utilities: \code{\link{ISOdatetime.j}},
\code{\link{ISOtimestring}},
\code{\link{irisws-timeutilities}},
\code{\link{timestring}}; \code{\link{check.query}},
\code{\link{iris.query}}, \code{\link{query.iris}};
\code{\link{constructor2.iriswadl}},
\code{\link{describe}}, \code{\link{describe.iriswadl}},
\code{\link{irisws-wadl}}, \code{\link{is.iriswadl}},
\code{\link{waddler}}, \code{\link{wadl}};
\code{\link{constructor}}, \code{\link{constructor2}},
\code{\link{constructor2.default}},
\code{\link{irisws-queryconstruction}},
\code{\link{params2queryparams}}
}

