% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{irisws-package}
\alias{irisws}
\alias{irisws-package}
\title{Access to IRIS Web Services within R}
\description{
This package can be used to obtain products from the
Data Management Center  of the
Incorporated Research Institutions in Seismology [1],
through their web-services API [2].

Currenty there is support for access to
\code{\link{Distance-Azimuth}} calculations,
xxx, ...
}
\details{
The code constructs queries according to
the requirements of the API, executes them, and
optionally saves and/or loads the results into
the R environment.
A good place to start is with the \code{\link{webservices}}
function.
}
\examples{
# see the available functions
webservices()

# see which webservices are actually supported
services()
}
\author{
A.J. Barbour
}
\references{
[1] \url{http://www.iris.edu/}

[2] \url{http://service.iris.edu/irisws}
}
\seealso{
\code{\link{irisws-webservices}} to see a list
of the webservice accessors currently available

\code{\link{constructor}} \emph{et al.} to construct queries

\code{\link{query.iris}} to query IRIS WS

Other Utilities: \code{\link{ISOdatetime.j}},
  \code{\link{ISOtimestring}},
  \code{\link{irisws-timeutilities}},
  \code{\link{timestring}}; \code{\link{check.query}},
  \code{\link{iris.query}}, \code{\link{query.iris}};
  \code{\link{constructor}}, \code{\link{constructor2}},
  \code{\link{constructor2.default}},
  \code{\link{irisws-queryconstruction}},
  \code{\link{params2queryparams}}; \code{\link{describe}},
  \code{\link{describe.iriswadl}},
  \code{\link{irisws-wadl}}, \code{\link{is.iriswadl}},
  \code{\link{parameters}},
  \code{\link{parameters.iriswadl}}, \code{\link{waddler}},
  \code{\link{wadl}}
}

