\docType{package}
\name{irisws-package}
\alias{irisws-package}
\title{Access to IRIS Web Services within R}
\description{
  Description
}
\details{
  Details
}
\author{
  A.J. Barbour
}
\references{
  \url{http://service.iris.edu/irisws}
}
\seealso{
  \code{\link{irisws-webservices}} to see a list of the
  webservice accessors currently available

  \code{\link{constructor}} \emph{et al.} to construct
  queries

  \code{\link{query.iris}} to query IRIS WS

  Other Utilities: \code{\link{check.query}},
  \code{\link{constructor}}, \code{\link{constructor2}},
  \code{\link{iris.query}},
  \code{\link{params2queryparams}},
  \code{\link{query.iris}}, \code{\link{timestring}}
}

