% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/webservices.R
\docType{package}
\name{irisws-webservices}
\alias{irisws-webservices}
\alias{irisws-webservices-package}
\alias{services}
\alias{webservice}
\alias{webservices}
\alias{ws}
\title{Summary of webservice accessors}
\usage{
webservices()

services()
}
\description{
Provides a list of all webservice
accessors  that are currently supported.
}
\details{
\code{\link{webservices}} lists the functions with
their arguments; whereas, \code{\link{services}}
lists the name of the associated web-service at
IRIS.
}
\examples{
# list all the webservice accessors:
webservices()

# just the names of webservices, as set by the IRIS API:
services()
}
\references{
\url{http://service.iris.edu/irisws}
}
\seealso{
\code{\link{irisws-package}}

Other WebServices: \code{\link{.llpair}},
  \code{\link{PS_time.distances}},
  \code{\link{PS_time.stations}}, \code{\link{traveltime}},
  \code{\link{traveltimes}}, \code{\link{tt}},
  \code{\link{ttDeg.ws}}, \code{\link{ttKm.ws}},
  \code{\link{ttStaSrc.ws}}, \code{\link{ws.traveltime}},
  \code{\link{ws.tt}}, \code{\link{ws.ttDeg}},
  \code{\link{ws.ttDistances}}, \code{\link{ws.ttKm}},
  \code{\link{ws.ttStaSrc}};
  \code{\link{Distance-Azimuth}}, \code{\link{distaz}},
  \code{\link{distaz.ws}}, \code{\link{ws.distaz}};
  \code{\link{flinn}}, \code{\link{flinnengdahl}},
  \code{\link{flinnengdahl.ws}},
  \code{\link{ws.flinnengdahl}}; \code{\link{resp}},
  \code{\link{resp.ws}}, \code{\link{ws.resp}};
  \code{\link{timeseries}}, \code{\link{timeseries.ws}},
  \code{\link{ws.timeseries}}
}

