% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webservices.R
\docType{package}
\name{irisws-webservices}
\alias{irisws-webservices}
\alias{irisws-webservices-package}
\alias{services}
\alias{webservice}
\alias{webservices}
\alias{ws}
\title{Summary of webservice accessors}
\usage{
webservices()

services()
}
\description{
Provides a list of all webservice
accessors  that are currently supported.
}
\details{
\code{\link{webservices}} lists the functions with
their arguments; whereas, \code{\link{services}}
lists the name of the associated web-service at
IRIS.
}
\examples{
# list all the webservice accessors:
webservices()

# just the names of webservices, as set by the IRIS API:
services()
}
\references{
\url{http://service.iris.edu/irisws}
}
\seealso{
\code{\link{irisws-package}}

Other WebServices: \code{\link{distaz}},
  \code{\link{flinnengdahl}}, \code{\link{resp}},
  \code{\link{timeseries}}, \code{\link{traveltime}}
}

