\docType{package}
\name{irisws-webservices}
\alias{irisws-webservices}
\alias{irisws-webservices-package}
\alias{webservices}
\title{Summary of webservice accessors}
\usage{
  webservices()
}
\description{
  Description
}
\details{
  Details
}
\examples{
# list all the webservice accessors:
webservices()
}
\references{
  \url{http://service.iris.edu/irisws}
}
\seealso{
  \code{\link{irisws-package}}

  Other WebServices: \code{\link{distaz}},
  \code{\link{distaz.ws}}, \code{\link{flinnengdahl}},
  \code{\link{flinnengdahl.ws}},
  \code{\link{PS_time.distances}},
  \code{\link{timeseries}}, \code{\link{timeseries.ws}},
  \code{\link{traveltime}}, \code{\link{ttDeg.ws}},
  \code{\link{ttKm.ws}}, \code{\link{ttStaSrc.ws}},
  \code{\link{ws.distaz}}, \code{\link{ws.flinnengdahl}},
  \code{\link{ws.timeseries}}, \code{\link{ws.ttDeg}},
  \code{\link{ws.ttDistances}}, \code{\link{ws.ttKm}},
  \code{\link{ws.ttStaSrc}}
}

