\name{resp}
\alias{resp}
\alias{resp.ws}
\alias{ws.resp}
\title{Access the 'resp' web service for instrument response information}
\usage{
ws.resp(network, station, location, channel, time = "", starttime = "",
  endtime = "", filename = NA, verbose = FALSE)

resp.ws(network, station, location, channel, time = "", starttime = "",
  endtime = "", filename = NA, verbose = FALSE)
}
\arguments{
  \item{network}{character; Regular network (e.g.,
  \code{'IU'}) or virtual network (e.g., \code{'_FDSN'})}

  \item{station}{charcter; Station code}

  \item{location}{character; Location code. Use \code{--}
  instead of spaces. Wildcards are accepted.}

  \item{channel}{character; Channel code. Wildcards are
  accepted.}

  \item{time}{character; Find the response for the given
  time. \code{time} cannot be used with \code{starttime} or
  \code{endtime}.  Using this can tend to produce}

  \item{starttime}{character; find response information
  from this point on}

  \item{endtime}{character; find response information up
  until this point}

  \item{filename}{\code{NA} for an auto-generated filename
  based on the inputs; \code{NULL} for a temporary
  filename; or, a character string of the user's choosing.
  See \strong{Details} for details about the auto-generated
  name.}

  \item{verbose}{logical; should messages be given?}
}
\value{
A list (invisibly) with the query string, and data from the
result
}
\description{
From [1]: \emph{The RESP webservice gives access to
instrument response information. The response information
is returned in the RESP format [2]. The service can return
responses for a channel, one or more channels under a
location, or one or more channels under a station.}
}
\details{
\subsection{Filename options (\code{filename=})}{ \itemize{
\item{\code{NA}}: An auto-generated file.
\item{\code{NULL}}: A temporary file.  \item{<character
string>}: Any desired name.  } Autogenerated files are of
the form:

\code{iriswsQ.<network>.<station>.<location>.<channel>.resp}

Temporary files are generated within
\code{\link{query.iris}}. }
}
\examples{
ws.resp("PB","B084","--","LDD")
try(ws.resp("PB","B084","--","LDD", time=timestring(2012,10,0,0,0)))
try(ws.resp("PB","B084","--","LDD", starttime=timestring(2012,10,0,0,0)))
}
\author{
AJ Barbour
}
\references{
[1] \url{http://service.iris.edu/irisws/resp/1/}

[2]
\url{http://www.iris.edu/KB/questions/69/What+is+a+RESP+file?}
}
\seealso{
Other WebServices: \code{\link{.llpair}},
\code{\link{PS_time.distances}},
\code{\link{PS_time.stations}}, \code{\link{traveltime}},
\code{\link{traveltimes}}, \code{\link{tt}},
\code{\link{ttDeg.ws}}, \code{\link{ttKm.ws}},
\code{\link{ttStaSrc.ws}}, \code{\link{ws.traveltime}},
\code{\link{ws.ttDeg}}, \code{\link{ws.ttDistances}},
\code{\link{ws.ttKm}}, \code{\link{ws.ttStaSrc}};
\code{\link{distaz}}, \code{\link{distaz.ws}},
\code{\link{ws.distaz}}; \code{\link{flinnengdahl}},
\code{\link{flinnengdahl.ws}},
\code{\link{ws.flinnengdahl}};
\code{\link{irisws-webservices}},
\code{\link{irisws-webservices-package}},
\code{\link{webservice}}, \code{\link{webservices}},
\code{\link{ws}}; \code{\link{timeseries}},
\code{\link{timeseries.ws}}, \code{\link{ws.timeseries}}
}

