\name{timeseries}
\alias{timeseries}
\alias{timeseries.ws}
\alias{ws.timeseries}
\title{Access to the 'timeseries' Web Service for obtaining continuous data}
\usage{
  ws.timeseries(network, station, location, channel,
    starttime, duration, endtime = NULL,
    output = c("sac.bin", "ascii", "plot", "ascii.values", "sac.asc", "miniseed", "audio"),
    filename = NA, endianness = c("auto", "little", "big"),
    verbose = TRUE, ...)

  timeseries.ws(network, station, location, channel,
    starttime, duration, endtime = NULL,
    output = c("sac.bin", "ascii", "plot", "ascii.values", "sac.asc", "miniseed", "audio"),
    filename = NA, endianness = c("auto", "little", "big"),
    verbose = TRUE, ...)
}
\arguments{
  \item{network}{character; the network code}

  \item{station}{character; the station code}

  \item{location}{character; the location code}

  \item{channel}{character; the channel code}

  \item{starttime}{character; the beginning of the record}

  \item{duration}{numeric; the length of the record, in
  seconds.  This will be ignored if \code{enddate} is not
  \code{NULL}.}

  \item{endtime}{character; the end of the record}

  \item{output}{character; the type of file to output to.
  See \strong{Details}}

  \item{filename}{\code{NA} for an auto-generated filename
  based on the inputs; \code{NULL} for a temporary
  filename; or, a character string of the user's choosing.
  See \strong{Details} for details about the auto-generated
  name.}

  \item{endianness}{character; specify the endianness of
  binary SAC output. \code{'auto'} uses the platform value,
  or \code{'little'} and \code{'big'} can be used to force
  a specific structure.}

  \item{verbose}{logical; should messages be given?}

  \item{...}{additional query parameters. Because
  \code{\link{constructor2}} is used, any bogus options are
  ignored.}
}
\value{
  A list (invisibly) with the query string, and data from
  the result
}
\description{
  From [1]: \emph{The ws-timeseries service provides access
  to individual channels of time series data for a
  specified time range. Requested segments may be up to 30
  days in length and optional signal processing may be
  applied to return data. The time series may be returned
  in a variety of formats.}
}
\details{
  The query is generated with \code{\link{constructor2}},
  and executed with \code{\link{query.iris}}, which thows
  errors based on \code{\link{check.query}}.

  \subsection{Output format options (\code{output=})}{
  \itemize{ \item{\code{'miniseed'}}: miniSEED
  \item{\code{'sac.asc'}}: SAC Alpha (ASCII)
  \item{\code{'sac.bin'}}: SAC binary (either big or little
  endian) \item{\code{'audio'}}: Audio (WAV format)
  \item{\code{'plot'}}: PNG plot
  \item{\code{='ascii.values'}}: ASCII (values only)
  \item{\code{='ascii'}}: ASCII (values, and datetimes) } }

  \subsection{Filename options (\code{filename=})}{
  \itemize{ \item{\code{NA}}: An auto-generated file of the
  form:
  \code{irisws.<network>.<station>.<location>.<channel>.<starttime>.<end>.<extension>}
  where \code{end} is either '\code{endtime}', or
  '\code{<duration>s}' and \code{extension} will depend on
  \code{output}.  \item{\code{NULL}}: A temporary file, as
  generated by \code{\link{query.iris}} \item{character
  string}: Any desired name.  } }

  \subsection{Signal processing options}{ \itemize{
  \item{high, low and band-pass filter (\code{})}
  \item{remove mean value (\code{})} \item{scaling by
  constant value (\code{})} \item{deconvolution of
  instrument response (with frequency limits and unit
  conversion) (\code{})} \item{differentiation and
  integration (\code{})} \item{decimation to lower sample
  rates (\code{})} } }
}
\examples{
\dontrun{
ws.timeseries("PB","B084","--","LDD","2013-10-01T00:00:00","100",output="plot")
ws.timeseries("PB","B084","--","LDD",timestring(2013,01,0,0,0,month=10),"100",output="plot")
}
}
\author{
  AJ Barbour
}
\references{
  [1] \url{http://service.iris.edu/irisws/timeseries/1/}
}
\seealso{
  \code{\link{timestring}} to make properly formatted time
  strings

  Other WebServices: \code{\link{distaz}},
  \code{\link{distaz.ws}}, \code{\link{traveltime}},
  \code{\link{ttDeg.ws}}, \code{\link{ttKm.ws}},
  \code{\link{ttStaSrc.ws}}, \code{\link{ws.distaz}},
  \code{\link{ws.ttDeg}}, \code{\link{ws.ttDistances}},
  \code{\link{ws.ttKm}}, \code{\link{ws.ttStaSrc}}
}

